+++ Code/Effects/PlayerEffects/IncreaseSpawnRateEffect.cs
@@ -1,4 +1,5 @@
-﻿using CrowdControlMod.Globals;
+﻿using System;
+using CrowdControlMod.Globals;
 using CrowdControlMod.ID;
 using CrowdControlMod.Utilities;
 using Terraria;
@@ -23,6 +24,7 @@ public sealed class IncreaseSpawnRateEffect : CrowdControlEffect
     public IncreaseSpawnRateEffect(int duration) : base(EffectID.IncreaseSpawnRate, duration, EffectSeverity.Negative)
     {
         CrowdControlNPC.EditSpawnRateHook += EditSpawnRate;
+        On_NPC.SpawnNPC += OnSpawnNPC;
     }
 
     #endregion
@@ -49,13 +51,58 @@ public sealed class IncreaseSpawnRateEffect : CrowdControlEffect
 
     private void EditSpawnRate(Player player, ref int spawnRate, ref int maxSpawns)
     {
-        if ((NetUtils.IsSinglePlayer && CrowdControlMod.GetInstance().IsSessionActive && IsActive) ||
-            (NetUtils.IsServer && IsActiveOnServer(player)))
+        if (!IsActiveFor(player))
         {
-            // Set the spawn rate if the effect is active for the player
-            spawnRate = (int)(spawnRate / Factor);
-            maxSpawns = (int)(maxSpawns * Factor);
+            return;
         }
+
+        // Set the spawn rate if the effect is active for the player
+        spawnRate = (int)(spawnRate / Factor);
+        maxSpawns = (int)(maxSpawns * Factor);
+    }
+
+    private void OnSpawnNPC(On_NPC.orig_SpawnNPC orig)
+    {
+        // Make it so the town is no longer safe, so that enemies spawn instead of just critters
+
+        Span<float> cachedValues = stackalloc float[Main.maxPlayers];
+        for (var i = 0; i < Main.maxPlayers; i++)
+        {
+            var player = Main.player[i];
+            if (!player.active)
+            {
+                continue;
+            }
+
+            cachedValues[i] = player.townNPCs;
+            if (!IsActiveFor(player))
+            {
+                continue;
+            }
+
+            // Trick the game into thinking the player isn't in a town
+            player.townNPCs = 0f;
+        }
+
+        // Naturally spawn an NPC
+        orig.Invoke();
+
+        // Restore cached values
+        for (var i = 0; i < Main.maxPlayers; i++)
+        {
+            var player = Main.player[i];
+            if (!player.active)
+            {
+                continue;
+            }
+
+            player.townNPCs = cachedValues[i];
+        }
+    }
+
+    private bool IsActiveFor(Player player)
+    {
+        return (NetUtils.IsSinglePlayer && CrowdControlMod.GetInstance().IsSessionActive && IsActive) || (NetUtils.IsServer && IsActiveOnServer(player));
     }
 
     #endregion