+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Changed "Reforge Item" effect to be able to reforge any valid item.
+- Changed "Increase Spawn Rate" effect to allow enemies to spawn in towns.
 - Fixed "Catch Critter" challenge not removing bug net from hand in some cases.
 
 ## [2.0.5] - 2023-09-20